Pending tasks are stored in a priority queue.
Worker threads pop the task on the top of the queue (if there is one), and then
check the schedule information for that task. If the current time is smaller that the
task's schedule then the thread waits until the task schedule is reached.

If a new task is queued the thread is waken up in order to execute the new task, this is
done because a thread should never block un future tasks.